+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/EnterPasswordScreen.java
@@ -72,7 +72,7 @@ public class EnterPasswordScreen extends VoiceChatScreenBase {
     @Override
     public boolean keyPressed(KeyEvent keyEvent) {
         if (keyEvent.isEscape()) {
-            minecraft.setScreen(null);
+            minecraft.setScreenAndShow(null);
             return true;
         }
         if (super.keyPressed(keyEvent)) {