+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/group/GroupScreen.java
@@ -91,7 +91,7 @@ public class GroupScreen extends VoiceChatScreenBase {
 
         leave = new ImageButton(guiLeft + xSize - buttonSize - 7, buttonY, LEAVE, button -> {
             ClientServerNetManager.sendToServer(new LeaveGroupPacket());
-            minecraft.setScreen(new JoinGroupScreen());
+            minecraft.setScreenAndShow(new JoinGroupScreen());
         });
         leave.setTooltip(Tooltip.create(LEAVE_GROUP));
         addRenderableWidget(leave);