+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/onboarding/ActivationOnboardingScreen.java
@@ -29,13 +29,13 @@ public class ActivationOnboardingScreen extends OnboardingScreenBase {
 
         Button ptt = Button.builder(Component.translatable("message.voicechat.onboarding.activation.ptt.name"), button -> {
             VoicechatClient.CLIENT_CONFIG.microphoneActivationType.set(MicrophoneActivationType.PTT).save();
-            minecraft.setScreen(new PttOnboardingScreen(this));
+            minecraft.setScreenAndShow(new PttOnboardingScreen(this));
         }).bounds(guiLeft, guiTop + contentHeight - BUTTON_HEIGHT * 2 - PADDING, contentWidth / 2 - PADDING / 2, BUTTON_HEIGHT).build();
         addRenderableWidget(ptt);
 
         Button voice = Button.builder(Component.translatable("message.voicechat.onboarding.activation.voice.name"), button -> {
             VoicechatClient.CLIENT_CONFIG.microphoneActivationType.set(MicrophoneActivationType.VOICE).save();
-            minecraft.setScreen(new VoiceActivationOnboardingScreen(this));
+            minecraft.setScreenAndShow(new VoiceActivationOnboardingScreen(this));
         }).bounds(guiLeft + contentWidth / 2 + PADDING / 2, guiTop + contentHeight - BUTTON_HEIGHT * 2 - PADDING, contentWidth / 2 - PADDING / 2, BUTTON_HEIGHT).build();
         addRenderableWidget(voice);