+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/onboarding/OnboardingManager.java
@@ -20,7 +20,7 @@ public class OnboardingManager {
     }
 
     public static void startOnboarding(@Nullable Screen parent) {
-        MC.setScreen(getOnboardingScreen(parent));
+        MC.setScreenAndShow(getOnboardingScreen(parent));
     }
 
     public static Screen getOnboardingScreen(@Nullable Screen parent) {
@@ -32,7 +32,7 @@ public class OnboardingManager {
         VoicechatClient.CLIENT_CONFIG.disabled.set(false).save();
         VoicechatClient.CLIENT_CONFIG.onboardingFinished.set(true).save();
         ClientManager.getPlayerStateManager().onFinishOnboarding();
-        MC.setScreen(null);
+        MC.setScreenAndShow(null);
     }
 
     public static void onConnecting() {