+++ fabric/changelog.md
@@ -1,3 +1,5 @@
 - Added access to the player in the voice host event
 - Prevent processing of UDP packets that are too large
 - Added rate limit config option
+- Fixed crash when opening ModMenu voice chat options screen without cloth config
+- Fixed GUI backgrounds