+++ fabric/changelog.md
@@ -1,1 +1,1 @@
-- Allow 26.1.x
+- Updated to 26w14a